chore: remove deprecated (#147)
Hard deprecation of items that were soft deprecated in v0.9

auto-generate vimdoc
